requirements.txt
ultralytics==8.2.0
kafka-python==2.0.2
opencv-python==4.8.1.78
numpy==1.24.3